[ 007 ] Software sample
There's now a working software - renderer sample for simple. It simply (pun intended) draws squares and circles on the screen, showing the capabilities of a renderer based on the window resolution. There's also now a function that allows user to "safely" get and modify the framebuffer data based on the x and y position (check 'pixel(int, int)' function)