[ 012 ] Mouse offset and scale
I'm trying to implement basic mouse scaling and offsetting. This could be useful on lower window resolutions, render textures, maybe with UI rendering and such. For now, it seems opengl sample works without an issue. But it appears that software sample segfaults on random occasions. I'm going to tackle it asap