[ 013 ] Segmentation Fault fix
I fixed a segmentation fault occuring in the software sample. It was a problem with the keyboard symbols exceeding the size of the array of keysyms (imagine: keysym equals to 65527 and the limit is 16384. Pretty simple yet easy to miss bug). Also, I fixed some compiler warnings and fixed the naming of monitor_height() function