[ 014 ] Window flags system
A beginning of a bigger system of dedicated flags for manipulating the current window state. window_flags() is a function that must be called before the window creation; it set's the flags that window will process. window_state() is a function that processes the window flags. It must be called after the window creation, because it needs a valid window and display connection (at least on the x11 implementation). The system is pretty bare-bones, it doesn't do any error checking. Gonna fix that now